Extract AO/AA/jitter multi-pass orchestration into FrameRenderPipeline
Move the inline GTAO off-screen pipeline (scene -> GTAO -> denoise ->
composite -> FXAA/SMAA -> temporal jitter) out of GUIView::drawScene into a
new qsys::FrameRenderPipeline class. The pipeline owns the render targets and
the fullscreen post-process primitive and exposes a molstar-style lifecycle
(setSize / render / dispose). The live view path and the off-screen exporter
(renderAOColorFrame) now share the same render() instead of duplicating the
pass chain.

GUIView keeps the view/camera concerns: projection (computeAoConstants,
setUpProjMat with sub-pixel jitter offset), jitter control-flow state,
needsContinuousRedraw/forceRedraw, the UI overlay and swapBuffers. The scene
geometry is injected as a callback so the pipeline never references View
internals.

Desktop (OpenGL) behavior is unchanged: the GL call sequence is preserved
verbatim, with the sole exception of clearing the scene target before the
scene callback (independent of the model matrix, so the output is identical).

Portability seams for the future WebGL2 port are isolated in the pipeline:
m_jitterSupported degrades the jitter path gracefully when float color
attachments are unavailable, blitDepthToDefault is a single gated call, and
the GTAO normal source already keys off RenderTarget::hasNormal().

Build: task build_libcuemol2 OK; task run_gtest 1130/1130 passed.
